Trait rustacuda::memory::SetDestination

pub trait SetDestination<O: ?Sized>: Sealed {
    fn set_u32(&mut self, value: u32) -> CudaResult<()>;

    fn set_u16(&mut self, value: u16) -> CudaResult<()>;

    fn set_u8(&mut self, value: u8) -> CudaResult<()>;
}

Sealed trait implemented by types which represent allocated memory on the device, which can be initialized from a value.

Required methods

fn set_u32(&mut self, value: u32) -> CudaResult<()>

Initializes or sets device memory with sequence of 32-bit values.

Errors

If a CUDA error occurs, return the error.

fn set_u16(&mut self, value: u16) -> CudaResult<()>

Initializes or sets device memory with sequence of 16-bit values.

Errors

If a CUDA error occurs, return the error.

fn set_u8(&mut self, value: u8) -> CudaResult<()>

Initializes or sets device memory with sequence of 8-bit values.

Errors

If a CUDA error occurs, return the error.

Loading content...