Module rustacuda::prelude

This module re-exports a number of commonly-used types for working with RustaCUDA.

This allows the user to use rustacuda::prelude::*; and have the most commonly-used types available quickly.

Re-exports

pub use crate::context::Context;

pub use crate::context::ContextFlags;

pub use crate::device::Device;

pub use crate::module::Module;

pub use crate::stream::Stream;

pub use crate::stream::StreamFlags;

pub use crate::CudaFlags;

Structs

DeviceBuffer
Fixed-size device-side buffer. Provides basic access to device memory.

UnifiedBuffer
Fixed-size buffer in unified memory.

Traits

CopyDestination
Sealed trait implemented by types which can be the source or destination when copying data to/from the device or from one device allocation to another.