Trait rustacuda::prelude::CopyDestination

pub trait CopyDestination<O: ?Sized>: Sealed {
    fn copy_from(&mut self, source: &O) -> CudaResult<()>;

    fn copy_to(&self, dest: &mut O) -> CudaResult<()>;
}

Sealed trait implemented by types which can be the source or destination when copying data to/from the device or from one device allocation to another.

Required methods

fn copy_from(&mut self, source: &O) -> CudaResult<()>

Copy data from source. source must be the same size as self.

Errors

If a CUDA error occurs, return the error.

fn copy_to(&self, dest: &mut O) -> CudaResult<()>

Copy data to dest. dest must be the same size as self.

Errors

If a CUDA error occurs, return the error.