Module rustacuda::error

Types for error handling

Error handling in CUDA:

RustaCUDA uses the CudaError enum to represent the errors returned by the CUDA API. It is important to note that nearly every function in CUDA (and therefore RustaCUDA) can fail. Even those functions which have no normal failure conditions can return errors related to previous asynchronous launches.

Enums

CudaError
Error enum which represents all the potential errors returned by the CUDA driver API.

Type Definitions

CudaResult
Result type for most CUDA functions.

DropResult
Special result type for drop functions which includes the un-dropped value with the error.