Module rustacuda::function

Functions and types for working with CUDA kernels.

Structs

BlockSize
Dimensions of a thread block, or the number of threads in a block.

Function
Handle to a global kernel function.

GridSize
Dimensions of a grid, or the number of thread blocks in a kernel launch.

Enums

FunctionAttribute
All supported function attributes for Function::get_attribute