Struct rustacuda::prelude::DeviceBuffer

pub struct DeviceBuffer<T> { /* fields omitted */ }

Fixed-size device-side buffer. Provides basic access to device memory.

Methods

impl<T> DeviceBuffer<T>

pub unsafe fn uninitialized(size: usize) -> CudaResult<Self>

Allocate a new device buffer large enough to hold size T's, but without initializing the contents.

Errors:

If the allocation fails, returns the error from CUDA. If size is large enough that size * mem::sizeof::<T>() overflows usize, then returns InvalidMemoryAllocation.

Safety:

The caller must ensure that the contents of the buffer are initialized before reading from the buffer.

Examples:

use rustacuda::memory::*;
let mut buffer = unsafe { DeviceBuffer::uninitialized(5).unwrap() };
buffer.copy_from(&[0u64, 1, 2, 3, 4]).unwrap();

pub unsafe fn zeroed(size: usize) -> CudaResult<Self>

Allocate a new device buffer large enough to hold size T's and fill the contents with zeroes (0u8).

Errors:

If the allocation fails, returns the error from CUDA. If size is large enough that size * mem::sizeof::<T>() overflows usize, then returns InvalidMemoryAllocation.

Safety:

The backing memory is zeroed, which may not be a valid bit-pattern for type T. The caller must ensure either that all-zeroes is a valid bit-pattern for type T or that the backing memory is set to a valid value before it is read.

Examples:

use rustacuda::memory::*;
let buffer = unsafe { DeviceBuffer::zeroed(5).unwrap() };
let mut host_values = [1u64, 2, 3, 4, 5];
buffer.copy_to(&mut host_values).unwrap();
assert_eq!([0u64, 0, 0, 0, 0], host_values);

pub unsafe fn from_raw_parts(
ptr: DevicePointer<T>,
capacity: usize
) -> DeviceBuffer<T>

Creates a DeviceBuffer<T> directly from the raw components of another device buffer.

Safety

This is highly unsafe, due to the number of invariants that aren't checked:

ptr needs to have been previously allocated via DeviceBuffer or cuda_malloc.

ptr's T needs to have the same size and alignment as it was allocated with.

capacity needs to be the capacity that the pointer was allocated with.

Violating these may cause problems like corrupting the CUDA driver's internal data structures.

The ownership of ptr is effectively transferred to the DeviceBuffer<T> which may then deallocate, reallocate or change the contents of memory pointed to by the pointer at will. Ensure that nothing else uses the pointer after calling this function.

Examples:

use std::mem;
use rustacuda::memory::*;

let mut buffer = DeviceBuffer::from_slice(&[0u64; 5]).unwrap();
let ptr = buffer.as_device_ptr();
let size = buffer.len();

mem::forget(buffer);

let buffer = unsafe { DeviceBuffer::from_raw_parts(ptr, size) };

pub fn drop(dev_buf: DeviceBuffer<T>) -> DropResult<DeviceBuffer<T>>

Destroy a DeviceBuffer, returning an error.

Deallocating device memory can return errors from previous asynchronous work. This function destroys the given buffer and returns the error and the un-destroyed buffer on failure.

Example:

use rustacuda::memory::*;
let x = DeviceBuffer::from_slice(&[10, 20, 30]).unwrap();
match DeviceBuffer::drop(x) {
    Ok(()) => println!("Successfully destroyed"),
    Err((e, buf)) => {
        println!("Failed to destroy buffer: {:?}", e);
        // Do something with buf
    },
}

impl<T: DeviceCopy> DeviceBuffer<T>

pub fn from_slice(slice: &[T]) -> CudaResult<Self>

Allocate a new device buffer of the same size as slice, initialized with a clone of the data in slice.

Errors:

If the allocation fails, returns the error from CUDA.

Examples:

use rustacuda::memory::*;
let values = [0u64; 5];
let mut buffer = DeviceBuffer::from_slice(&values).unwrap();